refactor: MyRecordPage에 탭 메뉴 도입을 위해 기존 내용을 RecordView 컴포넌트로 분리
